requirements.txt
Django==3.2.6
django-filter==2.2.0
djangorestframework==3.10.3
joblib==0.14.0
Markdown==3.1.1
numpy== 1.21.1
pandas==1.3.1
requests==2.26.0
scikit-learn==0.24.2